+++ composer.json
@@ -74,13 +74,13 @@
         "liip/imagine-bundle": "<1.9",
         "nelmio/api-doc-bundle": "<2.13.5 || >=4.0",
         "rackspace/php-opencloud": "<1.6",
-        "sonata-project/block-bundle": "<3.17 || >=4.0",
+        "sonata-project/block-bundle": "<4.0 || >=5.0",
         "sonata-project/classification-bundle": "<3.0 || >= 5.0",
         "sonata-project/core-bundle": "<3.20",
         "sonata-project/doctrine-orm-admin-bundle": "<3.0",
         "sonata-project/formatter-bundle": "<3.4",
         "sonata-project/notification-bundle": "<3.3",
-        "sonata-project/seo-bundle": "<2.5 || >=3.0",
+        "sonata-project/seo-bundle": "<3.0 || >=4.0",
         "symfony/monolog-bundle": "<2.4",
         "tilleuls/ckeditor-sonata-media-bundle": "<1.0"
     },
@@ -92,12 +92,12 @@
         "matthiasnoback/symfony-config-test": "^4.2",
         "matthiasnoback/symfony-dependency-injection-test": "^4.0",
         "nelmio/api-doc-bundle": "^2.13.5",
-        "sonata-project/admin-bundle": "^3.91",
-        "sonata-project/block-bundle": "^3.17",
+        "sonata-project/admin-bundle": "^3.93",
+        "sonata-project/block-bundle": "^4.0",
         "sonata-project/datagrid-bundle": "^3.0.1",
         "sonata-project/doctrine-orm-admin-bundle": "^3.14",
         "sonata-project/notification-bundle": "^3.12",
-        "sonata-project/seo-bundle": "^2.5",
+        "sonata-project/seo-bundle": "3.x-dev",
         "symfony/browser-kit": "^4.4 || ^5.2",
         "symfony/phpunit-bridge": "^5.2",
         "symfony/security-csrf": "^4.4 || ^5.2"
