+++ src/Block/BaseMediaBlockService.php
@@ -1,0 +1,172 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the Sonata Project package.
+ *
+ * (c) Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Sonata\MediaBundle\Block;
+
+use Doctrine\ORM\Mapping\ClassMetadataInfo;
+use Sonata\AdminBundle\Form\Type\ModelListType;
+use Sonata\BlockBundle\Block\BlockContextInterface;
+use Sonata\BlockBundle\Block\Service\AbstractBlockService;
+use Sonata\BlockBundle\Block\Service\EditableBlockService;
+use Sonata\BlockBundle\Form\Mapper\FormMapper;
+use Sonata\BlockBundle\Model\BlockInterface;
+use Sonata\Doctrine\Model\ManagerInterface;
+use Sonata\MediaBundle\Admin\BaseMediaAdmin;
+use Sonata\MediaBundle\Model\MediaInterface;
+use Sonata\MediaBundle\Provider\Pool;
+use Symfony\Component\DependencyInjection\ContainerInterface;
+use Symfony\Component\Form\FormBuilder;
+use Symfony\Component\HttpFoundation\Response;
+use Twig\Environment;
+
+/**
+ * @author Thomas Rabaix <thomas.rabaix@sonata-project.org>
+ */
+abstract class BaseMediaBlockService extends AbstractBlockService implements EditableBlockService
+{
+    /**
+     * @var BaseMediaAdmin
+     */
+    protected $mediaAdmin;
+
+    /**
+     * @var ManagerInterface
+     */
+    protected $mediaManager;
+
+    /**
+     * @var ContainerInterface
+     */
+    private $container;
+
+    public function __construct(
+        Environment $twig,
+        ContainerInterface $container,
+        ManagerInterface $mediaManager
+    ) {
+        parent::__construct($twig);
+
+        $this->mediaManager = $mediaManager;
+        $this->container = $container;
+    }
+
+    /**
+     * @return Pool
+     */
+    public function getMediaPool()
+    {
+        return $this->getMediaAdmin()->getPool();
+    }
+
+    /**
+     * @return BaseMediaAdmin
+     */
+    public function getMediaAdmin()
+    {
+        if (!$this->mediaAdmin) {
+            $this->mediaAdmin = $this->container->get('sonata.media.admin.media');
+        }
+
+        return $this->mediaAdmin;
+    }
+
+    public function execute(BlockContextInterface $blockContext, ?Response $response = null): Response
+    {
+        // make sure we have a valid format
+        $media = $blockContext->getBlock()->getSetting('mediaId');
+        if ($media instanceof MediaInterface) {
+            $choices = $this->getFormatChoices($media);
+
+            if (!\array_key_exists($blockContext->getSetting('format'), $choices)) {
+                $blockContext->setSetting('format', key($choices));
+            }
+        }
+
+        return $this->renderResponse($blockContext->getTemplate(), [
+            'media' => $blockContext->getSetting('mediaId'),
+            'block' => $blockContext->getBlock(),
+            'settings' => $blockContext->getSettings(),
+        ], $response);
+    }
+
+    public function load(BlockInterface $block): void
+    {
+        $media = $block->getSetting('mediaId', null);
+
+        if (\is_int($media)) {
+            $media = $this->mediaManager->findOneBy(['id' => $media]);
+        }
+
+        $block->setSetting('mediaId', $media);
+    }
+
+    public function prePersist(BlockInterface $block): void
+    {
+        $block->setSetting('mediaId', \is_object($block->getSetting('mediaId')) ? $block->getSetting('mediaId')->getId() : null);
+    }
+
+    public function preUpdate(BlockInterface $block): void
+    {
+        $block->setSetting('mediaId', \is_object($block->getSetting('mediaId')) ? $block->getSetting('mediaId')->getId() : null);
+    }
+
+    public function configureCreateForm(FormMapper $formMapper, BlockInterface $block): void
+    {
+        $this->configureEditForm($formMapper, $block);
+    }
+
+    /**
+     * @return array
+     */
+    protected function getFormatChoices(?MediaInterface $media = null)
+    {
+        $formatChoices = [];
+
+        if (!$media instanceof MediaInterface) {
+            return $formatChoices;
+        }
+
+        $formats = $this->getMediaPool()->getFormatNamesByContext($media->getContext());
+
+        foreach ($formats as $code => $format) {
+            $formatChoices[$code] = $code;
+        }
+
+        return $formatChoices;
+    }
+
+    /**
+     * @return FormBuilder
+     */
+    protected function getMediaBuilder(FormMapper $formMapper)
+    {
+        // simulate an association ...
+        $fieldDescription = $this->getMediaAdmin()->getModelManager()->getNewFieldDescriptionInstance($this->mediaAdmin->getClass(), 'media', [
+            'translation_domain' => 'SonataMediaBundle',
+        ]);
+        $fieldDescription->setAssociationAdmin($this->getMediaAdmin());
+        $fieldDescription->setAdmin($formMapper->getAdmin());
+        $fieldDescription->setOption('edit', 'list');
+        $fieldDescription->setAssociationMapping([
+            'fieldName' => 'media',
+            'type' => ClassMetadataInfo::MANY_TO_ONE,
+        ]);
+
+        return $formMapper->create('mediaId', ModelListType::class, [
+            'sonata_field_description' => $fieldDescription,
+            'class' => $this->getMediaAdmin()->getClass(),
+            'model_manager' => $this->getMediaAdmin()->getModelManager(),
+            'label' => 'form.label_media',
+        ]);
+    }
+}
