+++ src/Resources/config/block.xml
@@ -9,32 +9,25 @@
     <services>
         <service id="sonata.media.block.media" class="%sonata.media.block.media.class%" public="true">
             <tag name="sonata.block"/>
-            <argument>sonata.media.block.media</argument>
-            <argument type="service" id="sonata.templating"/>
+            <argument type="service" id="twig"/>
             <argument type="service" id="service_container"/>
             <argument type="service" id="sonata.media.manager.media"/>
         </service>
         <service id="sonata.media.block.feature_media" class="%sonata.media.block.feature_media.class%" public="true">
             <tag name="sonata.block"/>
             <argument type="service" id="twig"/>
-            <!-- NEXT_MAJOR: Remove "null" argument -->
-            <argument>null</argument>
             <argument type="service" id="service_container"/>
             <argument type="service" id="sonata.media.manager.media"/>
         </service>
         <service id="sonata.media.block.gallery" class="%sonata.media.block.gallery.class%" public="true">
             <tag name="sonata.block"/>
             <argument type="service" id="twig"/>
-            <!-- NEXT_MAJOR: Remove "null" argument -->
-            <argument>null</argument>
             <argument type="service" id="service_container"/>
             <argument type="service" id="sonata.media.manager.gallery"/>
         </service>
         <service id="sonata.media.block.gallery_list" class="%sonata.media.block.gallery_list.class%" public="true">
             <tag name="sonata.block"/>
             <argument type="service" id="twig"/>
-            <!-- NEXT_MAJOR: Remove "null" argument -->
-            <argument>null</argument>
             <argument type="service" id="sonata.media.manager.gallery"/>
             <argument type="service" id="sonata.media.pool"/>
         </service>
