+++ src/Resources/config/seo_block.xml
@@ -9,28 +9,19 @@
         <service id="sonata.media.block.breadcrumb_view" class="%sonata.media.block.breadcrumb_view.class%">
             <tag name="sonata.block" context="breadcrumb"/>
             <tag name="sonata.breadcrumb"/>
-            <argument>gallery_view</argument>
-            <argument>sonata.media.block.breadcrumb_view</argument>
-            <argument type="service" id="sonata.templating"/>
-            <argument type="service" id="knp_menu.menu_provider"/>
+            <argument type="service" id="twig"/>
             <argument type="service" id="knp_menu.factory"/>
         </service>
         <service id="sonata.media.block.breadcrumb_index" class="%sonata.media.block.breadcrumb_index.class%">
             <tag name="sonata.block" context="breadcrumb"/>
             <tag name="sonata.breadcrumb"/>
-            <argument>gallery_index</argument>
-            <argument>sonata.media.block.breadcrumb_view</argument>
-            <argument type="service" id="sonata.templating"/>
-            <argument type="service" id="knp_menu.menu_provider"/>
+            <argument type="service" id="twig"/>
             <argument type="service" id="knp_menu.factory"/>
         </service>
         <service id="sonata.media.block.breadcrumb_view_media" class="%sonata.media.block.breadcrumb_media.class%">
             <tag name="sonata.block" context="breadcrumb"/>
             <tag name="sonata.breadcrumb"/>
-            <argument>media_view</argument>
-            <argument>sonata.media.block.breadcrumb_view_media</argument>
-            <argument type="service" id="sonata.templating"/>
-            <argument type="service" id="knp_menu.menu_provider"/>
+            <argument type="service" id="twig"/>
             <argument type="service" id="knp_menu.factory"/>
         </service>
     </services>
